break reciters screen into separate components
